pyproject.toml
[tool.poetry]
name = "synql"
version = "0.1.0"
description = ""
authors = ["Victorien Elvinger <victorien@elvinger.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10.0"
sqlschm = "0.8.0"
pysqlite3-binary = "0.5.0"

[tool.poetry.dev-dependencies]
black = "22.12.0"
mypy = "0.991"
pylint = "^2.16.2"
pytest = "7.2.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
mypy_path = "./stubs/"
